Changed float32 to float64, fixes #89

+++ models/foodtruck.go
@@ -8,7 +8,7 @@ type JSONFoodTruck struct {
 	Location    [2]float64   `json:"location" bson:"location"`
 	Owner       string       `json:"owner" bson:"owner"`
 	Status      bool         `json:"status" bson:"status"`
-	AvgRating   float32      `json:"avgRating" bson:"avgRating"`
+	AvgRating   float64      `json:"avgRating" bson:"avgRating"`
 	Hours       [7][2]string `json:"hours" bson:"hours"`
 	Reviews     []string     `json:"reviews" bson:"reviews"`
 	Photos      []string     `json:"photos" bson:"photos"`
